requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
cycler==0.10.0
Django==3.1.3
django-bootstrap4==3.0.1
future==0.18.2
importlib-metadata==2.1.1
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.2.0
torch==1.6.0
torchvision==0.7.0
typing-extensions==3.10.0.0
zipp==3.5.0
